+++ tests/base.py
@@ -29,6 +29,7 @@ class ToistoTestCase(unittest.TestCase):
     def tearDown(self) -> None:
         """Clear the registries."""
         Concept.instances.clear()
+        Concept.capitonyms.clear()
         Concept.homographs.clear()
 
     @staticmethod